+++ spec/board_loop_spec.rb
@@ -1,7 +1,6 @@
 # spec/board_loop_spec.rb
 
 require_relative '../lib/game_logic'
-require_relative '../bin/main'
 
 RSpec.describe BoardLoop do
   describe '#board_loop' do
