+++ spec/board_moves_spec.rb
@@ -1,14 +1,13 @@
 # spec/board_moves_spec.rb
 
 require_relative '../lib/game_logic'
-require_relative '../bin/main'
 
 RSpec.describe BoardMoves do
   describe '#player_moves1' do
-    let(:board_moves) { BoardMoves.new.player_moves1(board) }
+    let(:board_moves) { BoardMoves.new }
     let(:board) { MainBoard.new.main_board }
-    it 'should have class of BoardResults' do
-      expect(board_moves).to eql(nil)
+    it 'should have class of BoardMoves' do
+      expect(board_moves.class).to eql(BoardMoves)
     end
   end
 end
