+++ spec/main_board_spec.rb
@@ -1,6 +1,5 @@
 # spec/main_board_spec.rb
 
-require_relative '../bin/main'
 require_relative '../lib/game_logic'
 
 RSpec.describe MainBoard do
