+++ spec/user_spec.rb
@@ -1,6 +1,5 @@
 # spec/user_spec.rb
 
-require_relative '../bin/main'
 require_relative '../lib/game_logic'
 
 RSpec.describe User1 do
